Project restructure and merge http-server into lobby
This is a forward looking update that creates a number of new sub-projects, not yet all implemented.
This gives us a placeholder for moving files around as part of future updates.

Important in this change the http-server capability is broken, and grafted onto the lobby project.
Next iterations will work to launch the http-server when the lobby is launched.

Sub-project summary:
- lobby-client-java: this is a java socket client, existing code would be refactored into this project, largely from Messengers
- lobby-client-http: this is the http-client project renamed and is largely unchanged
- java-sockets: forward looking sub-project that will contain common java-socket code
- game-headed-client-java: forward looking sub-project to have java client code specific for interacting with headless game server.
- lobby-db-code: forward looking sub-project that will contain the DB layer of the lobby.
- lobby-engine: this layer is meant to help us keep network transport code split from logic code. It is a bit forward looking and
    ideally large sections of 'lobby' would be merged into this sub-projectg, as well as any other lobby-specific code remaining in game-core.
