+++ game-core/build.gradle
@@ -8,7 +8,7 @@ ext {
 
 dependencies {
     implementation project(':java-extras')
-    implementation project(':http-client')
+    implementation project(':lobby-client-http')
     implementation project(':swing-lib')
     implementation 'org.snakeyaml:snakeyaml-engine:1.0'
     implementation 'com.googlecode.soundlibs:jlayer:1.0.1.4'
