--- http-client/README.md
@@ -1,4 +1,0 @@
-## Overview
-
-Provides an API for communicating with an http-server. Needs a host URI
-to be initialized.
