+++ lobby-client-http/src/main/java/org/triplea/http/client/HttpClient.java
