+++ lobby-client-http/src/main/java/org/triplea/http/client/ServiceResponse.java
