+++ lobby-client-http/src/main/java/org/triplea/http/client/error/report/ErrorReportClientFactory.java
