+++ lobby-client-http/src/main/java/org/triplea/http/client/error/report/create/ErrorReport.java
