+++ lobby-client-http/src/main/java/org/triplea/http/client/github/issues/GithubIssueClient.java
