+++ lobby-client-http/src/main/java/org/triplea/http/client/throttle/MessageNotSentException.java
