+++ lobby-client-http/src/test/java/org/triplea/http/client/ServiceResponseTest.java
