+++ lobby-client-http/src/test/java/org/triplea/http/client/error/report/ErrorReportClientTest.java
