+++ lobby-client-http/src/test/java/org/triplea/http/client/error/report/ErrorReportResponseTest.java
