+++ lobby-engine/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation project(':java-extras')
+    implementation project(':lobby-client-http')
+}
