+++ lobby-engine/src/main/java/org/triplea/server/reporting/error/ErrorReportRequest.java
