+++ lobby-engine/src/main/java/org/triplea/server/reporting/error/ErrorReportResponseAdapter.java
