+++ lobby-engine/src/test/java/org/triplea/server/reporting/error/ErrorReportResponseAdapterTest.java
