+++ lobby/src/main/java/org/triplea/server/EnvironmentVariable.java
