+++ lobby/src/main/java/org/triplea/server/http/spark/controller/ErrorReportController.java
