+++ settings.gradle
@@ -2,11 +2,15 @@ rootProject.name='triplea'
 include 'game-core'
 include 'game-headed'
 include 'game-headless'
-include 'http-client'
-include 'http-server'
+include 'game-headless-client-java'
 include 'infrastructure'
 include 'java-extras'
+include 'java-sockets'
 include 'lobby'
+include 'lobby-client-http'
+include 'lobby-client-java'
 include 'lobby-db'
+include 'lobby-db-code'
+include 'lobby-engine'
 include 'swing-lib'
 include 'test-common'
